Abstract some methods in Argon & Bcrypt hashers. (#24474)
